Moved auth forms to different folder
