+++ SmartSaver/forms/FormMain.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Windows.Forms;
+using ePiggy.forms.auth;
 using ePiggy.Forms.Finances.Budget;
 using ePiggy.Forms.Finances.Goals;
 using ePiggy.Forms.Reports;
