+++ SmartSaver/forms/auth/FormChangePass.cs
@@ -1,36 +1,30 @@
 ﻿using System;
 using System.Windows.Forms;
 using ePiggy.Authentication;
+using ePiggy.Forms;
 using ePiggy.Utilities;
 
-namespace ePiggy.Forms
+namespace ePiggy.forms.auth
 {
-    public partial class FormRegister : Form
+    public partial class FormChangePass : Form
     {
+        private readonly string _email;
         public Handler DataHandler { get; }
-
-
-        public FormRegister(Handler dataHandler)
+        public FormChangePass(string email, Handler dataHandler)
         {
             InitializeComponent();
-            DataHandler = dataHandler;
             errorMessage.Text = "";
-        }
-
-        
-        private void backToLogin_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            FormUtilities.ChangeForm(this, new FormLogIn(DataHandler));
+            _email = email;
+            DataHandler = dataHandler;
         }
 
         private void registerButton_Click(object sender, EventArgs e)
         {
-            var email = emailInput.Text;
-            var pass = passwordInput1.Text;
-            var passConfirm = passwordInput2.Text;
-            if (!InputValidator.RegisterValidation(this, email, pass, passConfirm)) return;
-            UserAuth.Registration(email: email, pass: pass);
-            FormUtilities.ChangeForm(this, new FormMain(DataHandler));
+            var pass = password.Text;
+            var passconfirm = confirmPassword.Text;
+            if (!InputValidator.IsValidPasswordConfirm(this, pass, passconfirm)) return;
+            UserAuth.ChangePassword(_email, pass);
+            FormUtilities.ChangeForm(this, new FormLogIn(DataHandler));
         }
 
         public void ChangeErrorText(int id)
@@ -38,9 +32,6 @@ namespace ePiggy.Forms
             string text;
             switch (id)
             {
-                case 0:
-                    text = "Provided e-mail is not valid!";
-                    break;
                 case 1:
                     text = "Password must be at least 8 characters long";
                     break;
@@ -57,6 +48,4 @@ namespace ePiggy.Forms
             errorMessage.Text = text;
         }
     }
-
-        
 }
