+++ SmartSaver/forms/auth/FormRegister.resx
