+++ SmartSaver/utilities/InputValidator.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using ePiggy.forms;
 using ePiggy.Forms;
+using ePiggy.forms.auth;
 
 namespace ePiggy.Utilities
 {
